// src/rect/types/Rect.ts
/**
 * A type representing a rectangle on a 2D plane.
 */
export type Rect = {
  /**
   * The top bound.
   */
  top: number

  /**
   * The right bound.
   */
  right: number

  /**
   * The bottom bound.
   */
  bottom: number

  /**
   * The left bound.
   */
  left: number

  /**
   * The `width` value.
   */
  width: number

  /**
   * The `height` value.
   */
  height: number
}